Reverse test order to make things simpler
Instead of walking the list of tests all the way to the end for each
new test (square run-time), we now construct our linked list of tests
by prepending new tests at the start of the list (constant run-time).

This also shaves 3 lines off the code.

This makes tests behave differently, now tests will be run
bottom-to-top of the order that they are defined in a file. This should
not be an issue for independent unit tests. If it is, we never promised
anything about the order so client code will need to change.