Run tests in order of appearance in file
The last commit was not a good idea, chasing your test failures
backwards through a file (from bottom to top) feels bad.

We now have tests in a nice order and an efficient linked list. The
linked list was made more efficient with some more code, we now
remember both the start and end of the list. This allows easy append
AND easy traversal from the start.